Geo.Location (geo v0.1.0)

Ecto schema which represents a location

Each location entry is stored as a row in locations table. Schema uses PostgreSQL specific types for few fields. ip_address is stored as INET type and coordinate is stored as Point

Specific types are used for these fields instead of more simple types to allow type specific queries.

Location is stored with ip_address as primary key. And all other fields are optional

Types

t()

@type t() :: %Geo.Location{
  __meta__: term(),
  city: String.t(),
  coordinate: Geo.Coordinate.t(),
  country: String.t(),
  country_code: String.t(),
  inserted_at: DateTime.t(),
  ip_address: :inet.ip_address(),
  mystery_value: String.t(),
  updated_at: DateTime.t()
}

Struct which represents location data

Functions

changeset(location, params \\ %{})

@spec changeset(t(), map()) :: Ecto.Changeset.t()

Creates changeset for location struct using params

params is map of containing changes. Changes are normalised, cast and validated. If there is any error, error will be added to the changeset. See Ecto.Changeset docs for more information.

ip_address IPv4 or IPv6 IP address. It can be a string or a tuple (Erlang representation). Multiple types of IP address notations are supported. Such as short notation 127.1, 127. All are finally normalised to Erlang representation

coordinate can be a tuple of {latitude, longitude} where numbers can be represented as string or float. OR it can be Geo.Coordinate.t()

country_code must a string of 2 characters. value is normalised to upper-case

prepare_params(params)

@spec prepare_params(map()) :: {:ok, map()} | {:error, Ecto.Changeset.t()}

Prepare params for raw SQL operation such as insert_all and update_all

Params are validated along with auto-generated fields.

Returns {:ok, map} when params is valid, {:error, changeset} when params is invalid